+++ RePair/Coder/heap.h
@@ -34,6 +34,7 @@ Chile. Blanco Encalada 2120, Santiago, Chile. gnavarro@dcc.uchile.cl
 #define HEAPINCLUDED
 
 #include <stdlib.h>
+#include <cstdint>
 #include "basics.h"
 #include "arrayg.h"
 #include "records.h"
@@ -41,26 +42,26 @@ Chile. Blanco Encalada 2120, Santiago, Chile. gnavarro@dcc.uchile.cl
 static const int PRNH = 0;
 
 typedef struct 
-  { int freq;
-    int elems; // a pointer within freq array
-    int larger,smaller; // pointers within ff array
+  { uint64_t freq;
+    int64_t elems; // a pointer within freq array
+    int64_t larger,smaller; // pointers within ff array
   } Thfreq;
 
 typedef struct 
-  { int id;
-    int prev,next; // actually pointers within freq array
-    int fnode; // ptr to its freq node (ptr to ff)
+  { int64_t id;
+    int64_t prev,next; // actually pointers within freq array
+    uint64_t fnode; // ptr to its freq node (ptr to ff)
   } Thnode;
 
 typedef struct
   { Thnode *freq; // space for all frequent nodes is preallocated, sqrt(u)
-    int freef; // ptr to free list in freq
+    uint64_t freef; // ptr to free list in freq
     Thfreq *ff; // space for all frequencies of frequent nodes prealloc idem
-    int freeff; // ptr to free list in ff
-    int smallest,largest; // list of frequent ones (ptrs in ff)
+    uint64_t freeff; // ptr to free list in ff
+    int64_t smallest,largest; // list of frequent ones (ptrs in ff)
     Tarray *infreq; // vectors for infrequent ones
-    int sqrtu;
-    int max;  // max freq heap used
+    uint64_t sqrtu;
+    uint64_t max;  // max freq heap used
     Trarray *Rec; // records
   } Theap;
 
@@ -70,23 +71,23 @@ public:
 	// creates new empty heap
 	// 0<factor<1: occupancy factor
 	// sqrt(u)*max(minsize,n/factor) integers
-	static Theap createHeap (int u, Trarray *Rec, float factor, int minsize);
+	static Theap createHeap (uint64_t u, Trarray *Rec, float factor, uint64_t minsize);
 	// destroys H
 	static void destroyHeap (Theap *H);
 	// inc freq of pair Rec[id]
-	static void incFreq (Theap *H, int id);
+	static void incFreq (Theap *H, int64_t id);
 	// dec freq of pair Rec[id]
-	static void decFreq (Theap *H, int id);
+	static void decFreq (Theap *H, int64_t id);
 	// with freq 1
-	static void insertHeap (Theap *H, int id);
+	static void insertHeap (Theap *H, int64_t id);
 
-	static int extractMax (Theap *H);
+	static int64_t extractMax (Theap *H);
 	// remove elems with freq 1
 	static void purgeHeap (Theap *H); 
 	// repositions pair
-	static void heapRepos (Theap *H, int id);
+	static void heapRepos (Theap *H, int64_t id);
 
-	static void move (Tarray A, int i, int j, Trecord *rec);
+	static void move (Tarray A, int64_t i, int64_t j, Trecord *rec);
 	static void prnH (Theap *H);
 };
 #endif
